gke kubetest: Start migration
This migrates a handful of jobs to the new GKE deployer. There are a
couple of --deployer=bash cases now, too, where I've identified stuff
that still needs to be added. (I'm missing a generic create arg
pass-through, which I meant to add along the way.)

Along the way: Let cluster names be chosen automatically, give
projects to boskos to allocate.