Expose resource metrics using labels in boskos
Using one metric name for boskos resources allows for simpler queries
that translate across resource types and states. For instance, instead
of a query for `boskos_myType_myState` the query would be for
`boskos_resources{type="myType",state="myState"}`. Furthermore, queries
can now also be done across states for a type, or across types for a
state. For instance, to see the states for a specific type:

    sum(boskos_resources{type="myThings"}) by (state)

This allows for simpler dashboard creation as well. The current behavior
of producing a "0" count for type/state combinations that do not have
any record in Boskos but are requested via flag is preserved.

Metrics serving is moved to `/metrics` on `:9090` as convention
dictates.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>