boskos: Periodically update dynamic resources
Handling of the lifecycles for dynamic resources is moved to a separate
function, UpdateAllDynamicResources, which is called periodically
(configurable using --dynamic-resource-update-period).

This new function queries for all existing DynamicResourceLifeCycles
and dynamic resources and ensures that the specified number of resources
exist, and that any expired resources are deleted.

When a DynamicResourceLifeCycle is removed from the configuration, the
MinCount and MaxCount for that DRLC in storage are set to 0, which
allows for graceful removal of any associated dynamic resources. Once
all resources have been deleted, the DRLC will be removed.