Boskos: Move config loading into controller
Currently we reconcile the config on file update or the dynamic part
periodically. This doesn't catch all cases, for example if non-dynamic
resources need to be deleted but are currently in use, this will only
happen when the config file gets changed. Additionally, its slow as it
happens only periodicially.

This change moves that into a simplistic controller that reacts to
config file, resource object and dynamic resource object changes.