Boskos: Manage a label that denotes leader to allow HA setups
This change adds a small controller to boskos that will make it manage a
label on pods. The label will be set on the pod that has the leader
lease and removed on all other pods. This in turn allows to have
multiple replicas of boskos running by using a service that targets said
label.