Update CRDs to apiextensions.k8s.io/v1
WIP because:
- needs to be tested against kubernetes <1.22
- needs upgrade testing

The biggest change here is that the schema is now required. I wrote this
by looking at the types in crd/ and common/. Any field that was a map
got a `x-kubernetes-preserve-unknown-fields: true`

The most controversial change is probably adding the api-approved
annotation, to allow us to keep using the .k8s.io group. This should
mean that updating the CRD doesn't require migrating any of the
resources. As prior art, I'm looking at the fact that prow.k8s.io
did the same thing for its ProwJob CRD.

This is controversial because these CRD's aren't really part of core
kubernetes. They should be moved to the x-k8s.io group. However it's not
immediately clear to me what it would take to have boskos support
transparently migrating resources from CRD's define in one group to
another.

This has been tested by running the example deployment against a kind
cluster running kubernetes 1.22